requirements.txt
boto==2.49.0
boto3==1.9.176
botocore==1.12.176
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.2
django-extensions==2.1.9
django-heroku==0.3.1
docutils==0.14
gensim==3.7.3
gunicorn==19.9.0
h5py==2.9.0
heroku==0.1.4
idna==2.8
imagekit==1.0.6
jmespath==0.9.4
JPype1==0.7.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
konlpy==0.5.1
numpy==1.16.4
pandas==0.24.2
psycopg2==2.8.3
python-dateutil==2.5.0
python-decouple==3.1
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
s3transfer==0.2.1
scipy==1.3.0
six==1.12.0
smart-open==1.8.4
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.2
